+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+Next release
+============
+
+### New features
+
+* [#236](https://github.com/statnett/Talk2PowerSystem_PM/issues/236): Update response from the `__about` endpoint to serve the required data for the components information page
+
+###################
+
 1.1.0rc1
 ============